+++ lib/src/math/collectionutils.dart
@@ -1,3 +1,18 @@
+/*
+Starfruit, a set of Dart utility libraries.
+Copyright (C) 2019 Aditya Kishore
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU Lesser General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Lesser General Public License for more details.
+*/
+
 import 'dart:math';
 
 StarCollectionsUtils get cUtils => StarCollectionsUtils();