+++ lib/src/math/stats/pairedstatsutils.dart
@@ -10,10 +10,11 @@ the Free Software Foundation, either version 3 of the License, or
 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
+GNU Lesser General Public License for more details.
 */
 
 import 'statsutils.dart';
+import 'package:linalg/linalg.dart';
 import 'dart:math';
 
 ///A class that takes in XY data and can perform some algebraic utilities
@@ -28,21 +29,19 @@ class StarStatsXY {
     xy = input;
   }
 
-  ///Get the given data
+  ///Get the given data.
   Map<num,num> get baseMap => xy;
 
-  ///Get the x coordinates of the given data
+  ///Get the x coordinates of the given data.
   List<num> get x => xy.keys.toList();
 
-  ///Get the y coordinates of the given data
+  ///Get the y coordinates of the given data.
   List<num> get y => xy.values.toList();
 
   ///Calculates the correlation coefficient.
   ///This is the strength of the linear relationship of given points.
   ///
-  ///(Also known as r)
-  ///
-  ///The larger |r| is, the stronger the correlation
+  ///(Also known as r) The larger abs(r) is, the stronger the correlation.
   num get corCoefficient {
     StarStats xStats = StarStats(x);
     StarStats yStats = StarStats(y);
@@ -58,19 +57,30 @@ class StarStatsXY {
     return r;
   }
 
+  ///Calculates the coefficient of determination.
+  ///This is the strength of the data fit for a given model.
+  ///
+  ///(Also known as r^2). The larger r^2, the stronger the data fit for the model.
+  ///
+  ///The adjusted determination coefficient is usually more accurate.
+  num get detCoefficient {
+    return pow(corCoefficient,2);
+  }
+
   ///Calculates the adjusted coefficient of determination.
   ///This is the strength of the data fit for a given model.
   ///
-  ///Takes the degree of the model (Quadratic => 2, Linear => 1)
+  ///(Also known as adjusted r^2). The larger adjusted r^2, the stronger the data fit for the model.
   ///
-  ///(Also known as adjusted r^2)
-  ///The larger r^2, the stronger the data fit for the model
-  num detCoefficient (int degree) {
+  ///More accurate than normal determination coefficient.
+  num get adjDetCoefficient {
+    int n = x.length;
     num orig =  pow(corCoefficient, 2);
-    return 1 - ((1-orig)*(x.length -1)) / (x.length - degree - 1);
+    return 1 - (1-orig)*((n-1)/(n-2));
   }
 
-  ///Calculate linear regression of a set of points: Returns a list of form y = mx + b, with linearReg[0] being m and linearReg[1] being b.
+  ///Calculate linear regression of a set of points: Returns a list of form y =
+  ///mx + b, with linearReg[0] being m and linearReg[1] being b.
   List<num> get linReg {
     List<num> slopevalue = [];
     StarStats xStats = StarStats(x);
@@ -96,20 +106,28 @@ class StarStatsXY {
     return slopevalue;
   }
 
-  static double sumW2 (List<num> x, int xdeg, List<num> y, int ydeg) {
-    num returner = 0;
-    for (var i = 0; i<x.length; i++) {
-      returner += pow(x[i], xdeg) + pow(y[i], ydeg);
+  ///Calculate quadratic regression of a set of points: Returns a list of form y =
+  ///ax^2 + bx + c, with quadReg[0] being m, quadReg[1] being b, and quadreg[2] being c.
+  List<num> get quadReg {
+    List<double> yy = [];
+    List<double> xx = [];
+    for (var numb in y) {
+      yy.add(numb.toDouble());
     }
-    return returner.toDouble();
-  }
-
-  static double sumW1 (List<num> x, int xdeg) {
-    num returner = 0;
-    for (var i = 0; i<x.length; i++) {
-      returner += pow(x[i], xdeg);
+    for (var numb in x) {
+      xx.add(numb.toDouble());
     }
-    return returner.toDouble();
+    List<double> xcol1 = List.filled(xx.length, 1.0);
+    List<double> xcol3 = List(xx.length);
+    for (var i = 0; i <xx.length; i++) {
+      xcol3[i] = pow(xx[i],2);
+    }
+    Matrix Y = Matrix([yy]).transpose();
+    Matrix X = Matrix([xcol1, xx, xcol3]).transpose();
+    Matrix B = (X.transpose() * X).inverse() * X.transpose() * Y;
+    var c = B[0][0];
+    var b = B[1][0];
+    var a = B[2][0];
+    return [a,b,c];
   }
-
 }